Fix ignored failures of MQTT subscriptions: will now emit an error event
Identify failed mqtt subscriptions, add clear log messages, and emit a
suitable error event.

The important change is the newly added error treatment in the MQTT
subscribe call, where the returned granted argument might list topics
with qos=128 which means those topics have not been subscribed. This is
a serious problem on which the application code must be able to react.